First Commit
mary

This tool is designed to aid in the creation of data driven suitability
layers. The tool takes a reference layer and a variable layer as inputs.
The reference layer represents the features that you want to create
suitabilities for ("Single Family Residential"), and the variable layer
represents the variable that relates to the reference layer (Parks for
Recreation). Based on the average and standard deviation of the distance
each object in the reference layer is from the variable layer, a
euclidean distance away from the variable layer is reclassified.